requirements.txt
# IRL-Based Intention Recognition - Python Dependencies

# Core Deep Learning & RL
torch>=1.9.0
stable-baselines3>=1.7.0

# Imitation Learning
imitation>=0.4.0

# Reinforcement Learning Environments
gymnasium>=0.28.0
pettingzoo>=1.22.0

# Game Rendering
pygame>=2.1.0

# Experiment Tracking
wandb>=0.12.0

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=0.19.0

# Utilities
numpy>=1.21.0

# Optional: For development
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
